+++ src/Commands/2_admin/prune.js
@@ -1,6 +1,6 @@
 const Command = require('../../Structures/Command.js');
 
-const { Permissions } = require('discord.js');
+const { PermissionsBitField } = require('discord.js');
 const { bot: { prefix, ownerID }, emoji: { success, warning }, response: { missingArguments, invalidPermissions, invalidNumber } } = require('../../../config/config.json');
 
 module.exports = new Command({
@@ -9,7 +9,7 @@ module.exports = new Command({
     syntax: 'prune <amount>',
 	description: 'Deletes the amount of messages send by the bot and the commands used to invoke the bot. Requires Manage Messages permission.',
 	async run(message, args, client) {
-		if (!message.member.permissions.has(Permissions.FLAGS.MANAGE_MESSAGES) && message.author.id != ownerID) return message.channel.send(`${warning} ${invalidPermissions} (Manage Messages)`);
+		if (!message.member.permissions.has(PermissionsBitField.Flags.ManageMessages) && message.author.id != ownerID) return message.channel.send(`${warning} ${invalidPermissions} (Manage Messages)`);
         if (!args[0]) return message.channel.send(`${warning} ${missingArguments}`);
         if (isNaN(args[0])) return message.channel.send(`${warning} ${invalidNumber}`);
         if (args[0] > 100 || args[0] < 1) return message.channel.send(`${warning} Outside of number range!`);
@@ -22,10 +22,10 @@ module.exports = new Command({
         let messagesDeleted = 0;
         for (let i = 0; i < resultArray.length; i++) {
             if (resultArray[i].author.id == client.user.id) {
-                resultArray[i].delete();
+                if ( (await message.channel.messages.fetch({ limit: 1, cache: false, around: resultArray[i].id })).has(resultArray[i].id) ) resultArray[i].delete();
                 messagesDeleted++;
                 if (resultArray[i+1].content.startsWith(prefix)) {
-                    resultArray[i+1].delete();
+                    if ( (await message.channel.messages.fetch({ limit: 1, cache: false, around: resultArray[i+1].id })).has(resultArray[i+1].id) )resultArray[i+1].delete();
                     messagesDeleted++;
                 }
             }
@@ -33,9 +33,9 @@ module.exports = new Command({
         }
 
         const response = await message.channel.send(`${success} Deleting ${args[0]} messages`)
-        setTimeout(() => {
-            response.delete();
-            message.delete();
+        setTimeout(async () => {
+            if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) )response.delete();
+            if ( (await message.channel.messages.fetch({ limit: 1, cache: false, around: message.id })).has(message.id) ) message.delete();
         }, 3750);
 	}
 });
